Add metadata to registered initializers and processes (closes #7).
Will allow things like silentExit to be set on a per-process basis.

This commit also changes the default behavior of a process exiting. Before, any process exiting with a nil error would be ignored. Now, it will shut down the runner unless WithSilentExit() is supplied at the process registration time.